Add boxing for closures stored in the extended Kalman filter.
Boxing is necessary because each closure has a different type, thus
it would be impossible to construct the AdditiveExtendedKalmanFilter if such
boxing was not there. Reason for this is that measurement and transition
functions would always have different types, but bounds would require them
to have same type, and this could never work.