Intermediate commit towards dimension-safe operations via wrappers for ndarray.
In this commit some basic wrappers are made available for executing operations
on ndarrays. The goal of these wrappers is to ensure that if correct-dimensional
states, measurements and configuration matrices are passed in, then any operation
is guaranteed to suceed.

Maybe a push towards further safety, such as positive-definiteness and symmetric
matrices will be done if necessary